+++ examples/example-twitter-server/src/main/scala/com/twitter/finatra/sample/modules/QueueModule.scala
@@ -1,0 +1,11 @@
+package com.twitter.finatra.sample.modules
+
+import com.twitter.finatra.sample.Queue
+import com.twitter.inject.TwitterModule
+
+object QueueModule extends TwitterModule {
+
+  override def configure(): Unit = {
+    bindSingleton[Queue].toInstance(new Queue())
+  }
+}